backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
alembic
cryptography
itsdangerous
passlib[bcrypt]
starlette
pyjwt